% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-helpers.R
\docType{data}
\name{apple_weather_trademark}
\alias{apple_weather_trademark}
\title{The Apple Weather Required Trademark}
\format{
An object of class \code{character} of length 1.
}
\usage{
apple_weather_trademark
}
\description{
Typing '' is a pain, so this should help folks stay legal without frustration.
The required legal link is in all the return objects.
}
\details{
\verb{ Weather"}
}
\keyword{datasets}
