% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temps.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Convert from Celsius to Fahrenheit}
\usage{
celsius_to_fahrenheit(T.celsius, round = 2)
}
\arguments{
\item{T.celsius}{Numeric vector of temperatures in Celsius.}

\item{round}{An integer indicating the number of decimal places to
round the converted value.}
}
\value{
double
}
\description{
Convert from Celsius to Fahrenheit
}
