% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerations.R
\docType{data}
\name{condition_codes}
\alias{condition_codes}
\title{Condition Codes}
\format{
An object of class \code{character} of length 40.
}
\usage{
condition_codes
}
\description{
seen in \code{conditionCode} list objects
}
\details{
\itemize{
\item \code{Clear}
\item \code{Cloudy}
\item \code{Dust}
\item \code{Fog}
\item \code{Haze}
\item \code{MostlyClear}
\item \code{MostlyCloudy}
\item \code{PartlyCloudy}
\item \code{ScatteredThunderstorms}
\item \code{Smoke}
\item \code{Breezy}
\item \code{Windy}
\item \code{Drizzle}
\item \code{HeavyRain}
\item \code{Rain}
\item \code{Showers}
\item \code{Flurries}
\item \code{HeavySnow}
\item \code{MixedRainAndSleet}
\item \code{MixedRainAndSnow}
\item \code{MixedRainfall}
\item \code{MixedSnowAndSleet}
\item \code{ScatteredShowers}
\item \code{ScatteredSnowShowers}
\item \code{Sleet}
\item \code{Snow}
\item \code{SnowShowers}
\item \code{Blizzard}
\item \code{BlowingSnow}
\item \code{FreezingDrizzle}
\item \code{FreezingRain}
\item \code{Frigid}
\item \code{Hail}
\item \code{Hot}
\item \code{Hurricane}
\item \code{IsolatedThunderstorms}
\item \code{SevereThunderstorm}
\item \code{Thunderstorm}
\item \code{Tornado}
\item \code{TropicalStorm}
}
}
\keyword{datasets}
