% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates-from-ip.R
\name{coordinates_from_ip}
\alias{coordinates_from_ip}
\title{Get latitude and longitude coordinates from your/an IP address}
\usage{
coordinates_from_ip(ip = NULL)
}
\arguments{
\item{ip}{if \code{NULL} (the default) the caller's IP address will be used.}
}
\value{
named numeric vector
}
\description{
Uses \href{https://ipapi.co/api/}{ipapi} (1,000/day; 30K/month free, no API key required)
to obtain the lat/lng of the caller's IP address or a specified one
}
