% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_conditions.R
\name{current_conditions}
\alias{current_conditions}
\title{Function to display current conditions for a given weather data retrieval}
\usage{
current_conditions(wx = wx_tidy(wxkit_weather(43.2683199, -70.8635506)))
}
\arguments{
\item{wx}{an object retrieved by \code{\link[=wxkit_weather]{wxkit_weather()}} and pre-tidied with \code{\link[=wx_tidy]{wx_tidy()}}}
}
\value{
nothing (side effect is printing to stdout)
}
\description{
By default this uses an API call and is the weather for the package author's
location. You likely want to specify \code{wx}.
}
