% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily-forecast.R
\name{daily_forecast}
\alias{daily_forecast}
\title{Function to display daily forecast for a given weather data retrieval}
\usage{
daily_forecast(
  wx = wx_tidy(wxkit_weather(43.2683199, -70.8635506)),
  tzone,
  metric = FALSE,
  incl_legal = TRUE
)
}
\arguments{
\item{wx}{an object retrieved by \code{\link[=wxkit_weather]{wxkit_weather()}} and pre-tidied with \code{\link[=wx_tidy]{wx_tidy()}}}

\item{tzone}{if not specified, this function will use \code{\link[lutz:tz_lookup_coords]{lutz::tz_lookup_coords()}}
"fast" method on the \code{forecastHourly$metadata} \code{latitude} and \code{longitude} values
to guess the timezone.}

\item{metric}{if \code{TRUE} (default is \code{FALSE}) data is in metric.}

\item{incl_legal}{if \code{FALSE} (default is \code{TRUE}) do not include Apple's required legal bits}
}
\value{
nothing (side effect is printing to stdout)
}
\description{
By default this uses an API call and is the weather for the package author's
location. You likely want to specify \code{wx}.
}
