% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temps.R
\name{dewpoint_to_humidity}
\alias{dewpoint_to_humidity}
\alias{D2H}
\title{Calculate relative humidity}
\usage{
dewpoint_to_humidity(dp = NA, t = NA, temperature.metric = "fahrenheit")

D2H(dp = NA, t = NA, temperature.metric = "fahrenheit")
}
\arguments{
\item{dp}{Numeric vector of dew point temperatures.}

\item{t}{Numeric vector of air temperatures.}

\item{temperature.metric}{Character string indicating the temperature
metric of air temperature and dew point temperature. Possible values
are \code{fahrenheit} or \code{celsius}.}
}
\value{
double
}
\description{
\code{dewpoint.to.humidity} creates a numeric vector of relative humidity
  from numerical vectors of air temperature and dew point temperature.
}
\details{
Dew point temperature and temperature must be in the same
   metric (i.e., either both in Celsius or both in Fahrenheit). If
   necessary, use \code{\link[weathermetrics]{convert_temperature}} to convert before
   using this function.
}
\note{
Equations are from the source code for the US National Weather
    Service's
    \href{http://www.wpc.ncep.noaa.gov/html/heatindex.shtml}{online heat index calculator}.
}
\examples{
# Calculate relative humidity from air temperature and
# dew point temperature.

data(lyon)
lyon$RH <- dewpoint.to.humidity(t = lyon$TemperatureC,
                                dp = lyon$DewpointC,
                                temperature.metric = 'celsius')
lyon
}
\references{
National Weather Service Hydrometeorological Prediction
   Center Web Team. Heat Index Calculator. 30 Jan 2015.
   \url{http://www.wpc.ncep.noaa.gov/html/heatindex.shtml}.
   Accessed 18 Dec 2015.
}
\seealso{
\code{\link[weathermetrics]{humidity.to.dewpoint},
   \link[weathermetrics]{fahrenheit.to.celsius},
   \link[weathermetrics]{celsius.to.fahrenheit}},
   \link[weathermetrics]{convert_temperature}
}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu},
Roger Peng \email{rdpeng@gmail.com}
}
