% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temps.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\alias{F2C}
\title{Convert from Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(T.fahrenheit, round = 2)

F2C(T.fahrenheit, round = 2)
}
\arguments{
\item{T.fahrenheit}{Numeric vector of temperatures in Fahrenheit.}

\item{round}{An integer indicating the number of decimal places to
round the converted value.}
}
\value{
double
}
\description{
\code{fahrenheit.to.celsius} creates a numeric vector of temperatures in
   Celsius from a numeric vector of temperatures in Fahrenheit.
}
\examples{
# Convert from Fahrenheit to Celsius.
data(norfolk)
norfolk$TempC <- fahrenheit.to.celsius(norfolk$TemperatureF)
norfolk
}
\seealso{
\code{\link[weathermetrics]{celsius.to.fahrenheit}}
}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu},
Roger Peng \email{rdpeng@gmail.com}
}
