% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{iso8601}
\alias{iso8601}
\title{Format a Datetime object: ISO 8601}
\usage{
iso8601(anytime("2016-09-01 10:11:12.123456"))
}
\arguments{
\item{pt}{A \code{POSIXt} Datetime or a \code{Date} object}
}
\value{
A character object formatted according to ISO 8601
}
\description{
Format a Datetime object: ISO 8601
}
