% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meters_to_miles}
\alias{meters_to_miles}
\title{Convert meters to miles}
\usage{
meters_to_miles(meters, round = TRUE)
}
\arguments{
\item{meters}{numeric}

\item{round}{round the return value? Default: \code{TRUE}}
}
\value{
numeric
}
\description{
Convert meters to miles
}
