% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temps.R
\name{metric_to_inches}
\alias{metric_to_inches}
\alias{M2I}
\title{Convert between standard metric units of measure for precipitation to inches}
\usage{
metric_to_inches(metric, unit.from, round = 2)

M2I(metric, unit.from, round = 2)
}
\arguments{
\item{metric}{Numeric vector of precipitation (in millimeters or centimeters)}

\item{unit.from}{A character string with the current units of the
observations (i.e., the units from which you want to convert)}

\item{round}{An integer indicating the number of decimal places to
round the converted value.}
}
\value{
double
}
\description{
\code{metric_to_inches} creates a numeric vector of precipitation measures in
  inches from a numeric vector of precipitation in common metric units
  (millimeters or centimeters).
}
\note{
Equations are from the source code for the National Weather Service
   Forecast Office \url{http://www.srh.noaa.gov/ama/?n=conversions}
}
\examples{
data(loveland)
loveland$Precip.in <- metric_to_inches(loveland$Precip.mm,
                                       unit.from = "mm",
                                       round = 2)
loveland
}
\references{
\url{http://www.srh.noaa.gov/ama/?n=conversions}
}
\seealso{
\code{\link[weathermetrics]{inches_to_metric}}
}
\author{
Joshua Ferreri \email{joshua.m.ferreri@gmail.com},
Brooke Anderson \email{brooke.anderson@colostate.edu}
}
