% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{precip_type}
\alias{precip_type}
\title{Converts \code{precipitationType} string values to symbols}
\usage{
precip_type(precip, daylight)
}
\arguments{
\item{precip}{\code{precipitationType} character vector}

\item{daylight}{\code{daylight} string logical}
}
\value{
character vector
}
\description{
Converts \code{precipitationType} string values to symbols
}
