% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pressure_trend}
\alias{pressure_trend}
\title{Converts \code{pressureTrend} string values to symbols}
\usage{
pressure_trend(trend)
}
\arguments{
\item{trend}{\code{pressureTrend} character vector}
}
\value{
character vector
}
\description{
Converts \code{pressureTrend} string values to symbols
}
