% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerations.R
\name{readable_conditions}
\alias{readable_conditions}
\title{Turns Apple's smushed together condition code strings into proper words/phrases}
\usage{
readable_conditions(condition_code)
}
\arguments{
\item{condition_code}{character vector}
}
\value{
character vector
}
\description{
Turns Apple's smushed together condition code strings into proper words/phrases
}
