% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_rng}
\alias{seq_rng}
\title{Create a sequence from a value range}
\usage{
seq_rng(vals, by = 1, na.rm = TRUE, ...)
}
\arguments{
\item{vals}{vector of values}

\item{by}{See \code{\link[=seq]{seq()}}}

\item{na.rm}{remove \code{NA}'s? Default: \code{TRUE}}

\item{...}{passed to \code{\link[=seq]{seq()}} call}
}
\description{
Create a sequence from a value range
}
