% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uv_label}
\alias{uv_label}
\title{UV index to label}
\usage{
uv_label(uv_index, swatch = FALSE)
}
\arguments{
\item{uv_index}{uv index level}

\item{swatch}{if \code{TRUE} (default is \code{FALSE}) a color swatch emoji
will be returned instead of the text label. 🟩 = \code{Low}, 🟨 = \code{Moderate}, 🟧 = \code{High},
🟥 = \verb{Very High},🟪 = \code{Extreme}}
}
\value{
character vector
}
\description{
UV index to label
}
