% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherkit-package.R
\docType{package}
\name{weatherkit}
\alias{weatherkit}
\alias{weatherkit-package}
\title{Obtain Historical, Current, and Predictive Weather Data}
\description{
Apple's WeatherKit REST API (\url{https://developer.apple.com/documentation/weatherkit})
provides timely weather information including current conditions, minute precipitation, along
with hourly, and daily forecasts. It also provides severe weather alerts. Tools are provided to
obtain historical, current, and predictive weather data. Also includes a set of weather icons by
Erik Flowers.
}
\details{
\itemize{
\item Temperatures are in Celsius
\item Precipitation intensity is in millimeters per hour.
\item Sea level air pressure in millibars.
}
}
\note{
If your apps, web apps, or websites display any Apple weather data (other than weather
alerts or value-added services or products, as described below), you must clearly display
the Apple Weather trademark ( Weather), as well as the
\href{://weatherkit.apple.com/legal-attribution.html}{legal link to other data sources}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/weatherkit}
  \item Report bugs at \url{https://github.com/hrbrmstr/weatherkit/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
