% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_tz}
\alias{with_tz}
\title{Get date-time in a different time zone}
\usage{
with_tz(time, tzone = "")
}
\arguments{
\item{time}{a POSIXct, POSIXlt, Date, chron date-time object.}

\item{tzone}{a character string containing the time zone to convert to. R must
recognize the name contained in the string as a time zone on your system.}
}
\description{
Returns a date-time as it would appear in a different time zone. The actual moment
of time measured does not change, just the time zone it is measured in. with_tz
defaults to the Universal Coordinated time zone (UTC) when an unrecognized time
zone is inputted. See \code{\link[=Sys.timezone]{Sys.timezone()}} for more information on how R recognizes time
zones.
}
\note{
Lifted from lubridate internals.
}
