% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{wx_icons_path}
\alias{wx_icons_path}
\title{Returns the full path to the installed SVG weather icons}
\usage{
wx_icons_path()
}
\value{
character vector
}
\description{
Returns the full path to the installed SVG weather icons
}
\examples{
list.files(wx_icons_path())
}
