% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{wx_tidy}
\alias{wx_tidy}
\title{Tidy values in weather data}
\usage{
wx_tidy(wx)
}
\arguments{
\item{wx}{object returned by \code{\link[=wxkit_weather]{wxkit_weather()}}}
}
\description{
Tidy values in weather data
}
\examples{
wx_tidy(wxkit_weather(43.2683199, -70.8635506))
}
