% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{wxkit_alert}
\alias{wxkit_alert}
\title{Receive an active weather alert.}
\usage{
wxkit_alert(id, language = Sys.getenv("LANG"), ..., auth = wxkit_auth())
}
\arguments{
\item{id}{The unique identifier for the weather alert.}

\item{language}{The language tag to use for localizing responses. Defaults to \code{Sys.getenv("LANG")}.}

\item{...}{passed to calls to \code{{httr}} functions.}

\item{auth}{See \code{\link[=wxkit_auth]{wxkit_auth()}}}

\item{references}{\url{https://developer.apple.com/documentation/weatherkitrestapi/get_api_v1_weatheralert_language_id}}
}
\value{
See: \url{https://developer.apple.com/documentation/weatherkitrestapi/weatheralert}
}
\description{
Receive an active weather alert.
}
