% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherkit-auth.R
\name{wxkit_auth}
\alias{wxkit_auth}
\title{Authentication components necessary for creation of JSON Web Tokens}
\usage{
wxkit_auth(
  key_id = Sys.getenv("WEATHERKIT_KEY_ID"),
  service_id = Sys.getenv("WEATHERKIT_SERVICE_ID"),
  team_id = Sys.getenv("WEATHERKIT_TEAM_ID"),
  private_key_path = Sys.getenv("WEATHERKIT_KEY_PATH")
)
}
\arguments{
\item{key_id}{10-character key identifier}

\item{service_id}{service identifier}

\item{team_id}{Apple Developer account team idenfitier}

\item{private_key_path}{path to private signing key (will be \code{\link[=path.expand]{path.expand()}}ed )}
}
\description{
Helper function that retrieves necessary parameters from the environment
or supplied by direct parameters
}
\references{
\href{https://developer.apple.com/documentation/weatherkitrestapi/request_authentication_for_weatherkit_rest_api}{Request authentication for WeatherKit REST API}
}
