% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{wxkit_availability}
\alias{wxkit_availability}
\title{Determine the data sets available for the specified location.}
\usage{
wxkit_availability(latitude, longitude, country, ..., auth = wxkit_auth())
}
\arguments{
\item{latitude}{The latitude of the desired location. Minimum Value: -90. Maximum Value: 90.}

\item{longitude}{The longitude of the desired location. Minimum Value: -180. Maximum Value: 180.}

\item{country}{The \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO Alpha-2 country code} for the requested location. This parameter is necessary for air quality and weather alerts.}

\item{...}{passed to calls to \code{{httr}} functions.}

\item{auth}{See \code{\link[=wxkit_auth]{wxkit_auth()}}}

\item{references}{\url{https://developer.apple.com/documentation/weatherkitrestapi/get_api_v1_availability_latitude_longitude}}
}
\value{
list of \href{https://developer.apple.com/documentation/weatherkitrestapi/dataset}{values}.

See: \url{https://developer.apple.com/documentation/weatherkitrestapi/weather}
}
\description{
Determine the data sets available for the specified location.
}
\examples{
wxkit_availability(43.2683199, -70.8635506, "US")
}
