% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{wxkit_weather}
\alias{wxkit_weather}
\title{Obtain weather data for the specified location.}
\usage{
wxkit_weather(
  latitude,
  longitude,
  language = Sys.getenv("LANG"),
  timezone = Sys.getenv("TZ"),
  dataSets = c("currentWeather", "forecastDaily", "forecastHourly", "forecastNextHour",
    "weatherAlerts"),
  countryCode = NULL,
  currentAsOf = NULL,
  dailyEnd = NULL,
  dailyStart = NULL,
  hourlyEnd = NULL,
  hourlyStart = NULL,
  ...,
  auth = wxkit_auth()
)
}
\arguments{
\item{latitude}{The latitude of the desired location. Minimum Value: -90. Maximum Value: 90.}

\item{longitude}{The longitude of the desired location. Minimum Value: -180. Maximum Value: 180.}

\item{language}{The language tag to use for localizing responses. Defaults to \code{Sys.getenv("LANG")}}

\item{timezone}{The name of the timezone to use for rolling up weather forecasts into daily forecasts.
Defaults to \code{Sys.getenv("TZ")}.}

\item{dataSets}{A character vector of data sets to include in the response. Call and cache the response to
\code{\link[=wxkit_availability]{wxkit_availability()}} for a list of \href{https://developer.apple.com/documentation/weatherkitrestapi/dataset}{possible values}
for the location you are querying data for. Defaults to all possible values.}

\item{countryCode}{OPTIONAL. The \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO Alpha-2 country code} for the requested location. This parameter is necessary for air quality and weather alerts.}

\item{currentAsOf}{OPTIONAL. The time to obtain current conditions. Defaults to now.}

\item{dailyEnd}{OPTIONAL. The time to end the daily forecast. If this parameter is absent, daily forecasts run for 10 days.}

\item{dailyStart}{OPTIONAL. The time to start the daily forecast. If this parameter is absent, daily forecasts start on the current day.}

\item{hourlyEnd}{OPTIONAL. The time to end the hourly forecast. If this parameter is absent, hourly forecasts run 24 hours or the length of the daily forecast, whichever is longer.}

\item{hourlyStart}{OPTIONAL. The time to start the hourly forecast. If this parameter is absent, hourly forecasts start on the current hour.}

\item{...}{passed to calls to \code{{httr}} functions.}

\item{auth}{See \code{\link[=wxkit_auth]{wxkit_auth()}}}

\item{references}{\url{https://developer.apple.com/documentation/weatherkitrestapi/get_api_v1_weather_language_latitude_longitude}}
}
\description{
Obtain weather data for the specified location.
}
\examples{
wxkit_weather(43.2683199, -70.8635506)
}
