% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temps.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{Convert from Fahrenheit to Celsius}
\usage{
fahrenheit_to_celsius(T.fahrenheit, round = 2)
}
\arguments{
\item{T.fahrenheit}{Numeric vector of temperatures in Fahrenheit.}

\item{round}{An integer indicating the number of decimal places to
round the converted value.}
}
\value{
double
}
\description{
Convert from Fahrenheit to Celsius
}
