% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherkit-package.R
\docType{package}
\name{weatherkit}
\alias{weatherkit}
\alias{weatherkit-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/weatherkit}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/weatherkit/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
