% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherkit-package.R
\docType{package}
\name{weatherkit}
\alias{weatherkit}
\alias{weatherkit-package}
\title{Obtain Historical, Current, and Predictive Weather Data}
\description{
Apple's WeatherKit REST API (\url{https://developer.apple.com/documentation/weatherkit})
provides timely weather information including current conditions, minute precipitation, along
with hourly, and daily forecasts. It also provides severe weather alerts. Tools are provided to
obtain historical, current, and predictive weather data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/weatherkit}
  \item Report bugs at \url{https://github.com/hrbrmstr/weatherkit/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
